services/olap-worker/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
nats-py==2.8.0
duckdb==1.1.1
asyncpg==0.29.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
